requirements.txt
#
# Requirements for Vadetis
#
django==2.2.7
django-extensions==2.2.5
django-celery-results==1.2.0
django-picklefield==2.0
mysqlclient==1.4.6
pandas==0.25.3
numpy==1.17.4
django-model-utils==3.2.0
kombu==4.6.7
celery==4.4.0
scipy==1.3.3
cython==0.29.14
confusable-homoglyphs==3.2.0
scikit-learn==0.21.3
matplotlib==3.1.2
redis==3.3.11
django-allauth==0.40.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-datatables==0.5.1
markdown==3.1.1
django-filter==2.2.0
coreapi==2.3.3
pyyaml==5.1.2
